# contracts/openapi/safety_kernel.yaml
...
openapi: 3.1.0
info:
  title: Safety Kernel Safety Kernel
  version: 1.0.0
  description: |
    Authorization, attestation, and audit-chain service for sensitive
    operations across the Safety Kernel system. The kernel is the
    only component authorized to mint signed tokens for actions like
    `sio_run_cycles`, self-improvement cycle persistence, and approval decisions.

    Caller roles (set by middleware via `x-api-key`):
    - `worker` — primary caller; authorized to request tokens for
      self-improvement/SIO/training actions
    - `api` — REST API; restricted to governance-only actions per
      `is_api_action_allowed()` allowlist
    - `operator` — human approvals/rejections only
  contact:
    name: ARYA Labs Safety Kernel team
    url: https://github.com/ARYA-Labs-Public/unfireable-safety-kernel
  license:
    name: Proprietary
servers:
  - url: http://safety-kernel:9000
    description: Internal docker network (production)
  - url: http://localhost:9001
    description: Host port mapping (local development)
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        Pre-shared key set in `.env` per role:
        - `QORCH_KERNEL_API_KEY_WORKER`
        - `QORCH_KERNEL_API_KEY_API`
        - `QORCH_KERNEL_API_KEY_OPERATOR`
  schemas:
...
    AuthorizeRequest:
      type: object
      required: [action, run_id, subject, params_fingerprint]
      properties:
        action:
          type: string
          minLength: 1
          description: Sensitive action to authorize (e.g., `sio_run_cycles`)
        run_id:
          type: string
          minLength: 1
          description: Run identifier bound into the token
        subject:
          type: string
          minLength: 1
          description: Worker identity / subject requesting authorization
        params_fingerprint:
          type: string
          minLength: 1
          pattern: ^[0-9a-f]{64}$
          description: SHA-256 fingerprint of parameters (stable JSON)
        params:
          type: object
          additionalProperties: true
          nullable: true
          description: Optional params used for policy decisions (fingerprinted)
        ttl_s:
          type: integer
          minimum: 1
          maximum: 3600
          nullable: true
          description: Requested TTL seconds (kernel will clamp)
        metadata:
          type: object
          additionalProperties: true
          nullable: true
          description: Optional metadata for audit (not used in token binding)
    AuthorizeResponse:
      type: object
      required: [ok, token, token_sha256, claims]
      properties:
        ok:
          type: boolean
          enum: [true]
        token:
          type: string
          description: Ed25519-signed compact JWT token
        token_sha256:
          type: string
          pattern: ^[0-9a-f]{64}$
          description: SHA-256 of the token string (for audit)
        claims:
          type: object
          additionalProperties: true
          description: Decoded claims (action, run_id, subject, exp, etc.)
...
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: string
          description: Stable machine-readable error code
        reason:
          type: string
          description: Optional human-readable reason

    # ------------------------------------------------------------------
    # /policy/* (ADR-018, internal-ref) — module-load / exec / compile
    # authorization. Slice 1 freezes ONLY PolicyModuleAuthorize{Request,
    # AllowResponse, DenyResponse}. The /policy/module/register,
    # /policy/audit-event, /policy/module/{path}/status endpoints are
    # listed in `paths:` with a DRAFT description and 501-only
    # responses; slice 2 fixes their request/response schemas.
    # ------------------------------------------------------------------
    PolicyModuleAuthorizeRequest:
      type: object
      required: [event_kind, module_path, caller_subject, caller_run_id, event_fingerprint]
      properties:
        event_kind:
          type: string
          enum: [import, exec, compile]
          description: CPython audit-event class that triggered the call
        module_path:
          type: string
          minLength: 1
          description: |
            For `import`: the dotted module name (`pkg.sub.mod`).
            For `exec` / `compile`: the SHA-256 hex of the source string
            (the source itself is NOT sent over the wire).
        caller_subject:
          type: string
          minLength: 1
          description: |
            Worker identity (matches the `subject` field on
            `/kernel/v1/authorize`). Lets the kernel correlate this
            event to an in-flight action authorization.
        caller_run_id:
          type: string
          minLength: 1
          description: Run identifier bound into the resulting decision token
        event_fingerprint:
          type: string
          pattern: ^[0-9a-f]{64}$
          description: |
            SHA-256 of the canonicalized event payload
            (event_kind, module_path, caller_subject, caller_run_id).
            Bound into the signed decision so a recorded ALLOW cannot
            be replayed against a different event.
        expected_required_patterns:
          type: array
          items:
            type: string
          nullable: true
          description: |
            Optional regex patterns the caller expects this `module_path`
            to require. Slice 2 uses these for drift validation: if the
            registry's recorded patterns do not equal the caller's
            expected set, the kernel returns 412 Precondition Failed.
            Lets the caller detect registry / code skew before the
            operation runs. Reflects the Rust `ModuleAuthorizeRequest`
            type in `crates/domain/src/safety/policy/types.rs`; ADR-018
            §"OpenAPI delta" amendment documents the slice-2 semantics.
        metadata:
          type: object
          additionalProperties: true
          nullable: true
          description: Optional audit metadata (not bound into the token)
    PolicyModuleAuthorizeAllowResponse:
      type: object
      required: [ok, decision, token, token_sha256, claims]
      properties:
        ok:
          type: boolean
          enum: [true]
        decision:
          type: string
          enum: [allow]
        token:
          type: string
          description: Ed25519-signed compact JWT, same envelope as `/kernel/v1/authorize`
        token_sha256:
          type: string
          pattern: ^[0-9a-f]{64}$
        claims:
          type: object
          additionalProperties: true
          description: Decoded claims (event_kind, module_path, exp, etc.)
    PolicyModuleAuthorizeDenyResponse:
      type: object
      required: [ok, decision, token, token_sha256, claims, reason]
      properties:
        ok:
          type: boolean
          enum: [false]
        decision:
          type: string
          enum: [deny]
        token:
          type: string
          description: Ed25519-signed DENY decision (replayable for audit)
        token_sha256:
          type: string
          pattern: ^[0-9a-f]{64}$
        claims:
          type: object
          additionalProperties: true
        reason:
          type: string
          minLength: 1
          description: Machine-readable refusal reason (e.g., `module_not_registered`)
  responses:
    Unauthorized:
      description: Missing / invalid `x-api-key`
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: |
        Role-based forbidden. `caller_role_forbidden` (unknown role) or
        `caller_role_not_operator` (operator-only endpoint) or
        `api_action_forbidden` (API role attempting non-governance action).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /kernel/v1/authorize:
    post:
      operationId: postAuthorize
      summary: Authorize a sensitive action (signed token)
      description: |
        Returns a signed Ed25519 token authorizing the specified action
        for the specified subject. The token's `exp` claim is the kernel's
        clamped TTL; clients MUST honour it. The audit chain records
        the authorization decision before the token is returned.
      tags: [kernel]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizeRequest'
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Validation error (Pydantic / FastAPI default body)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /policy/module/authorize:
    post:
      operationId: policyModuleAuthorize
      summary: Authorize a module-load or code-execution event
      description: |
        Called by the in-process Python audit hook (or any language
        runtime guard) on `import`, `exec`, or `compile` events.
        Returns an Ed25519-signed ALLOW token on success, an Ed25519-
        signed DENY decision with a `reason` on policy refusal, or
        503 `KernelUnavailable` if the kernel cannot reach its
        decision backend.

        Caller role: `worker` (the audit-hook reference uses the
        worker API key — it is part of the live workload process).
        `api` role is forbidden here (api actions go through
        `/kernel/v1/authorize`).
      tags: [policy]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyModuleAuthorizeRequest'
      responses:
        '200':
          description: ALLOW — signed decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyModuleAuthorizeAllowResponse'
        '403':
          description: DENY — signed decision with `reason`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyModuleAuthorizeDenyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: KernelUnavailable — decision backend unreachable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...